Implement Grantable entry fees [BREAKING]
// BREAKING
This commit changes the return type of the getEntryFee() method in the
Arena interface (and ArenaImpl class).
// BREAKING

This commit converts the entry fee system to the new Grantable system.
As a result, it is now possible to use arbitrary permissions as a type
of "token" to enter an arena. The "tokens" could possibly be granted
as rewards for reaching certain waves in always-accessible arenas. It
is also possible to use potion effects as entry fees, which is maybe a
bit strange, but possible, nonetheless.